+++ HISTORY.md
@@ -1,6 +1,6 @@
 2.0.0-alpha2 2014-10-02
 -----------------------
-- CRUD API (insertOne, insertMany, updateOne, updateMany, removeOne, removeMany, bulkWrite, findAndDelete, findAndUpdate, findAndReplace)
+- CRUD API (insertOne, insertMany, updateOne, updateMany, removeOne, removeMany, bulkWrite, findOneAndDelete, findOneAndUpdate, findOneAndReplace)
 - Cluster Management Spec compatible
 
 2.0.0-alpha1 2014-09-08